+++ test/tasty/neg/src-2/ConsumeTraitParams_fail.scala
@@ -1,0 +1,5 @@
+package tastytest
+
+object ConsumeTraitParams {
+  def test = TraitParams.foo // TODO: we should support this
+}
