+++ test/tasty/neg/src-2/TestLogarithms_fail.scala
@@ -1,0 +1,12 @@
+package tastytest
+
+import Logarithms._
+
+object TestLogarithms {
+
+  val Some(l1) = Logarithm.of(2)
+  val Some(l2) = Logarithm.of(3)
+
+  val l3: Double = l1 - l2 // error
+
+}
